+++ MichelCluster/Algo/TruncatedQBoundary.h
@@ -48,7 +48,7 @@ namespace michel {
     
     /// setter function for max distance between min in dQds
     /// and max in dQ
-    void SetMaxDistanceTruncatedPeaks(int n) { _maxDistance = n; }
+    void SetMaxDistanceTruncatedPeaks(size_t n) { _maxDistance = n; }
 
     /// setter function for strength SINGLE dip in covariance must be
     /// 0.9 is default
@@ -56,22 +56,22 @@ namespace michel {
 
     /// setter function for window size for calculating covariance and slope
     /// this in principle (and currently is) different from truncated Q window size
-    void SetCovarianceWindowSize(int s)      { _covariance_window = s; }
+    void SetCovarianceWindowSize(size_t s)   { _covariance_window = s; }
 
     /// setter function for window size used to calculated truncated Q
-    void SetTruncatedQWindowSize(int s)      { _n_window_size     = s; }
+    void SetTruncatedQWindowSize(size_t s)   { _n_window_size     = s; }
 
     /// setter function for fraction of points to keep in truncated Q calc
     void SetPAbove(double p)                 { _p_above = p; }
 
     // setter function for truncated Q window cutoff, windows less than
     // this size have no truncation
-    void SetMinWindowSize(int w)             { _window_cutoff = w; }
+    void SetMinWindowSize(size_t w)          { _window_cutoff = w; }
 
     /// setter function for mitigating edge effect, this number of points on the
     /// edge set equal to truncatedQ[_edgefix], maybe 0 is best I don't know
     
-    void SetEdgeEffectFix(int e)             { _edgefix = e; }
+    void SetEdgeEffectFix(size_t e)          { _edgefix = e; }
     
   private:
 
@@ -79,17 +79,17 @@ namespace michel {
     /// min in trunc. dQdx and max in trunc. dQ
     /// if the min and max are more than this number of hits
     /// apart -> do not create a michel
-    int _maxDistance;
+    size_t _maxDistance;
     
     /// single dip in covariance strenght, must be below this value to
     /// be considered a dip
     double _covariance_dip_cutoff;
     
     /// Sliding window size for calculating covariance/slope window
-    int _covariance_window;
+    size_t _covariance_window;
 
     /// Sliding window size for calculating truncated Q window
-    int _n_window_size;
+    size_t _n_window_size;
     
     /// Fraction of points to keep in truncated Q calculation, for example
     /// a value of 0.25 means we keep 25% of the lowest points in the specified
@@ -98,11 +98,11 @@ namespace michel {
 
     /// Minimum window size to apply truncation, if window size is smaller than this
     /// then we keep all the points 
-    int  _window_cutoff;
+    size_t  _window_cutoff;
     
     /// Try to correct edge effect by setting this number of points equal
     /// on both edges
-    int _edgefix;
+    size_t _edgefix;
      
   };
 }
